Uninject RDD methods
- Since only one method was being injected for each `RDD`
  transformation, the 6 classes used to introduce 3 function calls was
  decided to be too much ceremony. Eugene's advice was helpful here.

- Instead, these 6 classes have been removed in favour of three
  function calls in a single `object` in the highest package.
