+++ src/main/scala/vectorpipe/osm/package.scala
@@ -1,8 +1,6 @@
 package vectorpipe
 
-import geotrellis.spark.SpatialKey
 import geotrellis.vector._
-import org.apache.spark.rdd._
 import vectorpipe.util.Tree
 
 // --- //
@@ -15,9 +13,5 @@ package object osm {
   type OSMLine = Feature[Line, Tree[ElementData]]
   type OSMPolygon = Feature[Polygon, Tree[ElementData]]
   type OSMMultiPoly = Feature[MultiPolygon, Tree[ElementData]]
-
-  implicit class elementRDDMethods(rdd: RDD[Element]) extends ElementRDD(rdd)
-  implicit class featureRDDMethods(rdd: RDD[OSMFeature]) extends OSMFeatureRDD(rdd)
-  implicit class tiledOSMRDDMethods(rdd: RDD[(SpatialKey, Iterable[OSMFeature])]) extends TiledOSMRDD(rdd)
 
 }
